refactor: move auth_routes to api directory and change the login endpoint
